requirements.txt
streamlit
google-generativeai
pytesseract
pillow
PyMuPDF
mysql-connector-python
matplotlib
